+++ src/frontend/components/text-fields/active-editable-text-field.svelte
