+++ src/frontend/components/text-fields/passive-editable-text-field.svelte
