+++ src/frontend/icons/meta-system-icon.svelte
