+++ src/frontend/views/hub/project-preview-card.svelte
